% Generated by roxygen2 (4.0.1): do not edit by hand
\name{translate}
\alias{translate}
\title{Homology translation for miRNAtap}
\usage{
translate(entrezes, from = "mmu", to = "rno", ...)
}
\arguments{
\item{entrezes}{data.frame with entrez Gene IDs and their scores}

\item{from}{origin species}

\item{to}{target species}

\item{...}{any optional arguments}
}
\value{
data.frame object with orthologous genes' entrez IDs and
corresponding scores
}
\description{
Homology translation for miRNAtap
}
\details{
Translates gene ID using homology information from homologene.
}
\examples{
mouse_genes <- data.frame(GeneID =
        c("15364", "56520", "57781", "58180", "18035", "239857"))
translate(mouse_genes, from='mmu', to='rno')
}
\author{
Maciej Pajak \email{m.pajak@sms.ed.ac.uk}
}

